requirements.txt
appdirs==1.4.3
certifi==2017.1.23
cycler==0.10.0
dask==0.14.0
decorator==4.0.11
funcsigs==1.0.2
functools32==3.2.3.post2
h5py==2.6.0
mahotas==1.4.3
matplotlib==2.0.0
mock==2.0.0
networkx==1.11
numpy==1.12.0
olefile==0.44
packaging==16.8
pandas==0.19.2
pbr==2.0.0
Pillow==4.0.0
protobuf==3.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
scikit-image==0.12.3
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
#tensorflow-gpu==1.0.1
toolz==0.8.2
